Migrated JWT to a separate crate.
