+++ crates/shared/src/constants/general.rs
@@ -1,8 +1,5 @@
 pub const AUTHOR_NAME: &str = "Breno RdV";
 
-pub const JWT_APP_NAME: &str = "jwt";
-pub const JWT_VERSION: &str = "1.0.0";
-
 pub const NSCV_APP_NAME: &str = "nscv";
 pub const NSCV_VERSION: &str = "1.0.0";
