+++ crates/tool-jwt/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "jwt"
+version = "1.0.0"
+edition = "2021"
+authors = ["Breno RdV"]
+description = "Decodes JWT tokens and extracts their public claims, headers, and payload information. Optionally copies claims to clipboard."
+repository = "https://github.com/brenordv/rusted-toolbox"
+
+[dependencies]
+shared = { path = "../shared" }
+clap = "4.5.45"
+anyhow = "1.0.99"
+colored = "2.2.0"
+jsonwebtoken = "9.3.1"
+serde_json = "1.0.142"
+chrono = "0.4.41"
+serde = { version = "1.0.219", features = ["derive"] }
+tracing = "0.1.41"
