+++ crates/tool-jwt/src/jwt_app.rs
@@ -1,10 +1,10 @@
-use crate::tools::jwt::models::{Claims, ExpirationStatus, TokenInfo};
 use anyhow::{anyhow, Result};
 use colored::Colorize;
 use jsonwebtoken::{decode, Algorithm, DecodingKey, Validation};
 use serde_json::{Map, Value};
 use shared::utils::copy_string_to_clipboard::copy_to_clipboard;
 use std::process;
+use crate::models::{Claims, ExpirationStatus, TokenInfo};
 
 /// Decodes JWT token without signature verification.
 ///
