--- crates/tools/src/tools/jwt/mod.rs
@@ -1,3 +1,0 @@
-pub mod cli_utils;
-pub mod jwt_app;
-pub mod models;
