+++ crates/tools/src/tools/mod.rs
@@ -1,4 +1,3 @@
-pub mod jwt;
 pub mod split;
 pub mod touch;
 pub mod ts;
